Laid out the framework for a better Resource implementation
Created the code for a better implementation of the Resource page.  This version of the page uses 'tags' on each resource, in order to allow a link to span multiple resources.  The information is stored in the YAML header.

NOTE: THIS STILL NEEDS TO BE IMPLEMENTED!!!  The code is there, but is commented out until the sites in the Resource list are moved to the new standard.